+++ src/components/Box/BoxStyles.tsx
@@ -1,4 +1,4 @@
-import styled from 'styled-components'
+import styled, { css, FlattenSimpleInterpolation } from 'styled-components'
 import {
 	background,
 	border,
@@ -11,8 +11,9 @@ import {
 
 import { marginCss, paddingCss } from '../common/Spacing/SpacingStyles'
 import { textAlignCss } from '../common/Text/TextStyles'
+import { CommonBoxProps } from './CommonBoxProps'
 
-export const StyledBox = styled.div`
+export const StyledBox = styled.div<CommonBoxProps>`
 	${layout}
 	${flexbox}
 	${grid}
@@ -23,4 +24,11 @@ export const StyledBox = styled.div`
 	${paddingCss}
 	${textAlignCss}
 	${border}
+
+	${({ gap }): FlattenSimpleInterpolation | undefined =>
+		gap
+			? css`
+					gap: ${gap};
+			  `
+			: undefined}
 `