+++ src/components/Box/CommonBoxProps.tsx
@@ -1,0 +1,5 @@
+import * as CSS from 'csstype'
+
+export interface CommonBoxProps {
+	gap?: CSS.Property.Gap
+}