+++ src/types/ColorTheme.tsx
@@ -7,11 +7,3 @@ export const ColorTheme = [
 	'neutral'
 ] as const
 export type ColorTheme = typeof ColorTheme[number]
-
-export const ButtonColorTheme = [
-	'primary',
-	'success',
-	'error',
-	'neutral'
-] as const
-export type ButtonColorTheme = typeof ButtonColorTheme[number]